# Conflicts:
#	.monkeycode/docs/2026-06-17-p10-1-security-performance-usability.md
#	.monkeycode/docs/2026-06-17-p10-2-risk-mitigation-plan.md
#	.monkeycode/docs/2026-06-17-p2-3-rustc-1.94-fix.md
#	.monkeycode/docs/2026-06-17-p2-4-ai-extend-api.md
#	.monkeycode/docs/2026-06-17-p2-4-ai-extend-user-manual.md
#	.monkeycode/docs/2026-06-17-p3-1-microservice-api.md
#	.monkeycode/docs/2026-06-17-p3-1-microservice-user-manual.md
#	.monkeycode/docs/2026-06-17-p3-2-websocket-api.md
#	.monkeycode/docs/2026-06-17-p3-2-websocket-user-manual.md
#	.monkeycode/docs/2026-06-17-p3-3-react-native-api.md
#	.monkeycode/docs/2026-06-17-p3-3-react-native-user-manual.md
#	.monkeycode/docs/2026-06-17-p3-4-data-warehouse-api.md
#	.monkeycode/docs/2026-06-17-p3-4-data-warehouse-user-manual.md
#	.monkeycode/docs/2026-06-17-p4-1-perf-optimization.md
#	.monkeycode/docs/2026-06-17-p4-2-security-hardening.md
#	.monkeycode/docs/2026-06-17-p4-3-monitoring.md
#	.monkeycode/docs/2026-06-17-p4-4-i18n-guide.md
#	.monkeycode/docs/2026-06-17-p4-5-coverage-report.md
#	.monkeycode/docs/2026-06-17-p4-7-chaos-scenarios.md
#	.monkeycode/docs/2026-06-17-p4-7-disaster-recovery.md
#	.monkeycode/docs/2026-06-17-p4-8-ops-manual.md
#	.monkeycode/docs/2026-06-17-p5-1-final-evaluation.md
#	.monkeycode/docs/2026-06-17-p7-1-final-delivery-certificate.md
#	.monkeycode/docs/2026-06-17-p8-1-comprehensive-health-assessment.md
#	.monkeycode/docs/2026-06-17-p9-1-unwrap-cleanup.md
#	.monkeycode/docs/2026-06-17-p9-2-split-services.md
#	.monkeycode/docs/2026-06-17-p9-5-test-coverage.md
#	.monkeycode/docs/2026-06-17-p9-6-opentelemetry.md
#	.monkeycode/docs/2026-06-17-p9-7-kafka-integration.md
#	.monkeycode/docs/2026-06-17-p9-8-elasticsearch-integration.md
#	.monkeycode/docs/api/README.md
#	.monkeycode/docs/api/openapi.yaml
#	.monkeycode/docs/chaos-test-scenarios.md
#	.monkeycode/docs/color-card-api.md
#	.monkeycode/docs/color-card-deployment-guide.md
#	.monkeycode/docs/color-card-user-manual.md
#	.monkeycode/docs/color-price-api.md
#	.monkeycode/docs/color-price-deployment-guide.md
#	.monkeycode/docs/color-price-user-manual.md
#	.monkeycode/docs/custom-order-api.md
#	.monkeycode/docs/custom-order-deployment-guide.md
#	.monkeycode/docs/custom-order-user-manual.md
#	.monkeycode/docs/db/README.md
#	.monkeycode/docs/db/SCHEMA.md
#	.monkeycode/docs/failover-deployment-guide.md
#	.monkeycode/docs/quotation-api.md
#	.monkeycode/docs/quotation-user-manual.md
#	.monkeycode/docs/releases/v2026.617.0001.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-color-card-plan.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-color-price-extension-plan.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-custom-order-plan.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-failover-isolation-plan.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-p2-2-n+1-list.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-p2-2-perf-baseline.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-p2-2-slow-query-list.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-sales-quotation-plan.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-wave4-p2-1-evaluation.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-wave4-p2-2-plan.md
#	.monkeycode/docs/superpowers/plans/2026-06-16-wave4-p2-3-plan.md
#	.monkeycode/docs/superpowers/plans/2026-06-17-p3-1-microservice.md
#	.monkeycode/docs/superpowers/plans/2026-06-17-p3-2-websocket.md
#	.monkeycode/docs/superpowers/plans/2026-06-17-p3-3-react-native.md
#	.monkeycode/docs/superpowers/plans/2026-06-17-p3-4-data-warehouse.md
#	.monkeycode/docs/superpowers/reports/2026-06-16-code-audit.md
#	.monkeycode/docs/superpowers/reports/2026-06-16-failover-design.md
#	.monkeycode/docs/superpowers/reports/2026-06-16-industry-features.md
#	.monkeycode/docs/superpowers/reports/2026-06-16-industry-standards.md
#	.monkeycode/docs/superpowers/reports/2026-06-16-project-score.md
#	.monkeycode/docs/superpowers/specs/2026-06-16-color-card-design.md
#	.monkeycode/docs/superpowers/specs/2026-06-16-color-price-extension-design.md
#	.monkeycode/docs/superpowers/specs/2026-06-16-custom-order-design.md
#	.monkeycode/docs/superpowers/specs/2026-06-16-failover-isolation-design.md
#	.monkeycode/docs/superpowers/specs/2026-06-16-sales-quotation-design.md
#	.monkeycode/docs/superpowers/specs/2026-06-16-wave4-p2-2-perf-design.md
#	.monkeycode/docs/superpowers/specs/2026-06-16-wave4-p2-3-v2table-design.md
#	.monkeycode/docs/superpowers/specs/2026-06-17-p3-1-microservice.md
#	.monkeycode/docs/superpowers/specs/2026-06-17-p3-2-websocket.md
#	.monkeycode/docs/superpowers/specs/2026-06-17-p3-3-react-native.md
#	.monkeycode/docs/superpowers/specs/2026-06-17-p3-4-data-warehouse.md
#	CHANGELOG.md
#	MEMORY.md

+++ backend/src/models/sales_quotation_item.rs
@@ -1,13 +1,6 @@
 #![allow(dead_code)]
 // TODO(tech-debt): 业务接入或重评估后逐项移除；rustc 1.94+ 编译时由编译器报告具体死代码位置。
 
-//! 销售报价单明细实体
-//!
-//! 行项目表：每个产品/色号对应一行报价。包含色号编码、潘通号、CNCS 号、规格、计量单位、数量、单价、金额、阶梯价 JSON、折扣等。
-//! 关联计划：[2026-06-17-p12-batch1-quotation-port-plan.md](../../../../../docs/superpowers/plans/2026-06-17-p12-batch1-quotation-port-plan.md) PR-1
-//!
-//! 字段类型适配 main 风格：i32 与主表保持一致。
-
 use rust_decimal::Decimal;
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
@@ -16,49 +9,30 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
 #[sea_orm(table_name = "sales_quotation_items")]
 pub struct Model {
-    /// 明细 ID（主键）
     #[sea_orm(primary_key)]
-    pub id: i32,
-    /// 报价单 ID（外键 sales_quotations.id，级联删除）
-    pub quotation_id: i32,
+    pub id: i64,
+    pub quotation_id: i64,
 
-    /// 产品 ID（外键 products.id）
-    pub product_id: i32,
-    /// 色号 ID（外键 product_colors.id）
-    pub color_id: Option<i32>,
-    /// 色号编码
+    pub product_id: i64,
+    pub color_id: Option<i64>,
     pub color_code: Option<String>,
-    /// 潘通色号
     pub pantone_code: Option<String>,
-    /// CNCS 色号
     pub cncs_code: Option<String>,
 
-    /// 规格
     pub specification: Option<String>,
-    /// 计量单位
     pub unit: String,
 
-    /// 数量
     pub quantity: Decimal,
-    /// 不含税单价
     pub unit_price: Decimal,
-    /// 含税单价
     pub unit_price_with_tax: Decimal,
-    /// 不含税金额
     pub amount: Decimal,
-    /// 含税金额
     pub amount_with_tax: Decimal,
 
-    /// 阶梯价 JSON 数据
     pub tier_pricing: Option<Json>,
-    /// 折扣率（%）
     pub discount_rate: Option<Decimal>,
-    /// 折扣金额
     pub discount_amount: Option<Decimal>,
 
-    /// 备注
     pub notes: Option<String>,
-    /// 排序号
     pub sequence: i32,
 }
